services/video_service/requirements.txt
moviepy>2
pip
piper-tts
playwright
pre-commit
pydantic
